refactor: 각 컴포넌트 폴더에 stories 추가
