+++ .eslintrc.json
@@ -5,13 +5,15 @@
     "node": true
   },
   "extends": ["airbnb", "plugin:prettier/recommended"],
-  "parser": "babel-eslint",
+  // "parser": "babel-eslint",
   "rules": {
     "no-unused-vars": "warn",
     "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
     "react/react-in-jsx-scope": "warn",
     "react/prop-types": "warn",
     "react/require-default-props": "warn",
-    "no-param-reassign": "warn"
+    "no-param-reassign": "warn",
+    "react/jsx-props-no-spreading": "off",
+    "react/function-component-definition": "off"
   }
 }
