+++ src/components/card/Card.stories.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Card from '../components/Card.jsx';
-import '../css/index.css';
-import '../css/App.css';
+import Card from './Card';
+import '../../css/index.css';
+import '../../css/App.css';
 
 export default {
   title: 'Card',
